Read the namelist file only if it exists
Read the namelist file, pioperf.nl, only if it exists. The user
should be able to provide all input via command line args
(the user can now specify couple of them via command line, the
support for the rest of the args is being added) and hence
may want to ignore the namelist file.